0.9.7 Nearly finish.. Still need debug for error.

+++ CLP-AdaptiveSolution/trunk/CLP-AdaptiveSolution/classPlacement.vb
@@ -104,7 +104,7 @@ Public Class Placement
     ''' <summary>
     ''' Get utilization box in space
     ''' </summary>
-    Public ReadOnly Property Compactness() As Integer
+    Public ReadOnly Property Compactness() As Single
         Get
             Return fCompactness
         End Get
